+++ magpylib/_src/fields/field_BH_loop.py
@@ -12,10 +12,10 @@ from magpylib._src.input_checks import check_field_input
 
 # ON INTERFACE
 def current_loop_field(
+    field: str,
+    observers: np.ndarray,
     current: np.ndarray,
     diameter: np.ndarray,
-    observer: np.ndarray,
-    field='B'
     ) -> np.ndarray:
     """Magnetic field of a circular (line) current loop.
 
@@ -23,19 +23,18 @@ def current_loop_field(
 
     Parameters
     ----------
+    field: str, default=`'B'`
+        If `field='B'` return B-field in units of [mT], if `field='H'` return H-field
+        in units of [kA/m].
+
+    observers: ndarray, shape (n,3)
+        Observer positions (x,y,z) in Cartesian coordinates in units of [mm].
+
     current: ndarray, shape (n,)
         Electrical current in units of [A].
 
     diameter: ndarray, shape (n,)
         Diameter of loop in units of [mm].
-
-    observer: ndarray, shape (n,3)
-        Observer positions (x,y,z) in Cartesian coordinates in units of [mm].
-
-    field: str, default=`'B'`
-        If `field='B'` return B-field in units of [mT], if `field='H'` return H-field
-        in units of [kA/m].
-
     Returns
     -------
     B-field or H-field: ndarray, shape (n,3)
@@ -76,7 +75,7 @@ def current_loop_field(
 
     bh = check_field_input(field, 'current_loop_field()')
 
-    r, phi, z = cart_to_cyl_coordinates(observer)
+    r, phi, z = cart_to_cyl_coordinates(observers)
     rad = np.abs(diameter/2)
     n = len(rad)