+++ magpylib/_src/fields/field_BH_sphere.py
@@ -8,10 +8,10 @@ from magpylib._src.input_checks import check_field_input
 
 
 def magnet_sphere_field(
+    field: str,
+    observers: np.ndarray,
     magnetization: np.ndarray,
     diameter: np.ndarray,
-    observer: np.ndarray,
-    field='B'
     )->np.ndarray:
     """Magnetic field of a homogeneously magnetized sphere.
 
@@ -19,18 +19,18 @@ def magnet_sphere_field(
 
     Parameters
     ----------
+    field: str, default=`'B'`
+        If `field='B'` return B-field in units of [mT], if `field='H'` return H-field
+        in units of [kA/m].
+
+    observers: ndarray, shape (n,3)
+        Observer positions (x,y,z) in Cartesian coordinates in units of [mm].
+
     magnetization: ndarray, shape (n,3)
         Homogeneous magnetization vector in units of [mT].
 
     diameter: ndarray, shape (n,3)
         Sphere diameter in units of [mm].
-
-    observer: ndarray, shape (n,3)
-        Observer positions (x,y,z) in Cartesian coordinates in units of [mm].
-
-    field: str, default=`'B'`
-        If `field='B'` return B-field in units of [mT], if `field='H'` return H-field
-        in units of [kA/m].
 
     Returns
     -------
@@ -62,7 +62,7 @@ def magnet_sphere_field(
 
     # all special cases r0=0 and mag=0 automatically covered
 
-    x, y, z = np.copy(observer.T)
+    x, y, z = np.copy(observers.T)
     r = np.sqrt(x**2+y**2+z**2)   # faster than np.linalg.norm
     r0 = abs(diameter)/2
 
@@ -73,7 +73,7 @@ def magnet_sphere_field(
     mask_out = (r>=r0)
 
     mag1 = magnetization[mask_out]
-    obs1 = observer[mask_out]
+    obs1 = observers[mask_out]
     r1 = r[mask_out]
     r01 = r0[mask_out]