+++ magpylib/_src/fields/field_wrap_BH_level2.py
@@ -50,7 +50,7 @@ def get_src_dict(group: list, n_pix: int, n_pp: int, poso: np.ndarray) -> dict:
     kwargs = {
         "source_type": src_type,
         "position": posv,
-        "observer": posov,
+        "observers": posov,
         "orientation": rotobj,
     }
 
@@ -81,12 +81,7 @@ def get_src_dict(group: list, n_pix: int, n_pp: int, poso: np.ndarray) -> dict:
         kwargs.update({"current": currv, "vertices": vert_list})
 
     elif src_type == "CustomSource":
-        kwargs.update(
-            {
-                "field_B_lambda": group[0].field_B_lambda,
-                "field_H_lambda": group[0].field_H_lambda,
-            }
-        )
+        kwargs.update(field_func=group[0].field_func)
 
     else:
         raise MagpylibInternalError("Bad source_type in get_src_dict")
@@ -133,7 +128,7 @@ def getBH_level2(sources, observers, **kwargs) -> np.ndarray:
 
     # CHECK AND FORMAT INPUT ---------------------------------------------------
     if isinstance(sources, str):
-        return getBH_dict_level2(source_type=sources, observer=observers, **kwargs)
+        return getBH_dict_level2(source_type=sources, observers=observers, **kwargs)
 
     # bad user inputs mixing getBH_dict kwargs with object oriented interface
     kwargs_check = kwargs.copy()
@@ -153,10 +148,9 @@ def getBH_level2(sources, observers, **kwargs) -> np.ndarray:
 
     # test if all source dimensions and excitations are initialized
     check_dimensions(sources)
-    check_excitations(sources)
+    check_excitations(sources, kwargs['field'])
 
-
-    # format observer inputs:
+    # format observers input:
     #   allow only bare sensor, collection, pos_vec or list thereof
     #   transform input into an ordered list of sensors (pos_vec->pixel)
     #   check if all pixel shapes are similar - or else if pixel_agg is given
@@ -230,9 +224,7 @@ def getBH_level2(sources, observers, **kwargs) -> np.ndarray:
     groups = {}
     for ind, src in enumerate(src_list):
         if src._object_type == "CustomSource":
-            group_key = (
-                src.field_B_lambda if kwargs["field"] == "B" else src.field_H_lambda
-            )
+            group_key = src.field_func
         else:
             group_key = src._object_type
         if group_key not in groups: